collision: Add ray/cylinder intersection test
Add a ray/cylinder intersection test. It first checks for intersection
with the end planes, which can fairly easily be checked to see if within
the circle of the end of the cylinder. Finally it cancels out the motion
of the ray relative to the cylinder and solves the quadratic equation
for the ray/circle intersection.

Signed-off-by: James Hogan <james@albanarts.com>